Move downloads teardown into makevers and update tests accordingly
Make sure to remove downloads after test_get_new
