+++ test/test_get_new.py
@@ -2,12 +2,12 @@ from __future__ import absolute_import, print_function
 
 from ..pyautoupdate.launcher import Launcher
 from .pytest_skipif import needinternet
-from .pytest_makevers import fixture_update_dir
+from .pytest_makevers import fixture_update_dir,create_update_dir
 
 import os
 
 @needinternet
-def test_check_get_new(fixture_update_dir):
+def test_check_get_new(fixture_update_dir,create_update_dir):
     """Test that gets new version from internet"""
     package=fixture_update_dir("0.0.1")
     launch = Launcher('filling up the boring replacements',
@@ -16,6 +16,7 @@ def test_check_get_new(fixture_update_dir):
     with open(os.path.abspath("downloads/extradir/blah.py"), "r") as file_code:
         file_text=file_code.read()
     assert "new version" in file_text
+    assert os.path.isdir("downloads")
 
 @needinternet
 def test_check_get_invalid_archive(fixture_update_dir):
@@ -26,4 +27,5 @@ def test_check_get_invalid_archive(fixture_update_dir):
                       newfiles="project.tar.gz")
     launch._get_new()
     assert os.path.isfile("project.tar.gz.dump")
+    assert not os.path.isdir("downloads")
     os.remove("project.tar.gz.dump")
